backend/{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["{{cookiecutter.maintainer}}"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = { version = "^0.79.0", extras = ["all"] }
sqlmodel = "^0.0.6"
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
Faker = "^13.11.1"
gunicorn = "^20.1.0"
google-cloud-logging = "^3.0.0"
cryptography = "^37.0.3"
SQLAlchemy = "<1.4.42"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
black = "^22.1"
flake8 = "^4.0.1"


[tool.black]
line-length = 99                                                              # override black's default line-length
exclude = '(\.git|\.mypy_cache|\.tox|venv|\.venv|_build|buck-out|build|dist)'
